fix:修复了微信小程序用户查询菜品列表时遗漏了口味信息的bug

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -50,7 +50,7 @@ public class DishController {
 
     @GetMapping("/{id}")
     @ApiOperation("根据id查询菜品")
-    public Result findDishById(@PathVariable int id){
+    public Result findDishById(@PathVariable long id){
         DishVO dish = dishService.findDishById(id);
         return Result.success(dish);
     }
